Introduce publisher specification and user permissions to book model
This commit introduces a set of publisher options to the Book model and updates the user permissions related to these options. Corresponding changes are also made in the view and template files to accommodate these modifications. A new utility file is created to handle permissions related to the model. Additionally, the commit includes clean-ups such as the deletion of unused http files.